fix: address feedback
