+++ src/sentry/codecov/endpoints/TestResults/test_results.py
@@ -1,5 +1,3 @@
-from enum import Enum
-
 from drf_spectacular.utils import extend_schema
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -13,32 +11,7 @@ from sentry.codecov.base import CodecovEndpoint
 from sentry.codecov.client import CodecovApiClient
 from sentry.codecov.endpoints.TestResults.query import query
 from sentry.codecov.endpoints.TestResults.serializers import TestResultSerializer
-
-
-class OrderingDirection(Enum):
-    DESC = "DESC"
-    ASC = "ASC"
-
-
-class OrderingParameter(Enum):
-    AVG_DURATION = "AVG_DURATION"
-    FLAKE_RATE = "FLAKE_RATE"
-    FAILURE_RATE = "FAILURE_RATE"
-    COMMITS_WHERE_FAIL = "COMMITS_WHERE_FAIL"
-    UPDATED_AT = "UPDATED_AT"
-
-
-class TestResultsFilterParameter(Enum):
-    FLAKY_TESTS = "FLAKY_TESTS"
-    FAILED_TESTS = "FAILED_TESTS"
-    SLOWEST_TESTS = "SLOWEST_TESTS"
-    SKIPPED_TESTS = "SKIPPED_TESTS"
-
-
-class MeasurementInterval(Enum):
-    INTERVAL_30_DAY = "INTERVAL_30_DAY"
-    INTERVAL_7_DAY = "INTERVAL_7_DAY"
-    INTERVAL_1_DAY = "INTERVAL_1_DAY"
+from sentry.codecov.enums import MeasurementInterval, OrderingDirection, OrderingParameter
 
 
 @extend_schema(tags=["Prevent"])
